mc_backend/mcserver/requirements.txt
alembic==1.1.0
aniso8601==7.0.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blinker==1.4
boto3==1.12.27
botocore==1.15.27
cachetools==4.0.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmarkgfm==0.4.2
conllu==1.4
coverage==4.5.4
cycler==0.10.0
decorator==4.4.0
dnspython==1.16.0
docutils==0.15.2
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
future==0.17.1
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphannis==0.23.1
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
idna-ssl==1.1.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
monotonic==1.5
multidict==4.5.2
MyCapytain==2.0.10
networkx==2.4
nose==1.3.7
numpy==1.18.2
Pillow==6.1.0
pkginfo==1.5.0.1
protobuf==3.11.3
psutil==5.6.3
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
python-editor==1.0.4
python-rapidjson==0.8.0
pytz==2019.2
rapidjson==1.0.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
readme-renderer==24.0
reportlab==3.5.23
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==1.10.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
tqdm==4.35.0
twine==1.13.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
Werkzeug==0.15.6
wrapt==1.11.2
xhtml2pdf==0.2.3
yarl==1.3.0